+ set-exclusive-or
this was just the version I thought of first. will redefine to use "not equal" at some point as that more clearly represents how an exclusive-or is defined. compare the nested or/not versions of xor (exclusive-or) and a==b (equals) under the "logic" section. note also that keywords are not implemented, as they will not help a beginner's understanding.